+ Updated to 3236.

+++ trunk/trunk_3052/src/game/SpellAuras.cpp
@@ -4213,7 +4213,7 @@ void Aura::EventPeriodicLeech(uint32 amount)
                 amount = dmg.full_damage - dmg.resisted_damage;
 		}
 
-		uint32 Amount = min( amount, m_target->GetUInt32Value( UNIT_FIELD_HEALTH ) );
+		uint32 Amount = (uint32)min( amount, m_target->GetUInt32Value( UNIT_FIELD_HEALTH ) );
 		uint32 newHealth = m_caster->GetUInt32Value(UNIT_FIELD_HEALTH) + Amount ;
 		
 		uint32 mh = m_caster->GetUInt32Value(UNIT_FIELD_MAXHEALTH);
@@ -4660,7 +4660,7 @@ void Aura::EventPeriodicManaLeech(uint32 amount)
 	if(m_target->isAlive() && m_caster->isAlive())
 	{
 	
-		int32 amt = min( amount, m_target->GetUInt32Value( UNIT_FIELD_POWER1 ) );
+		int32 amt = (int32)min( amount, m_target->GetUInt32Value( UNIT_FIELD_POWER1 ) );
 		uint32 cm = m_caster->GetUInt32Value(UNIT_FIELD_POWER1)+amt;
 		uint32 mm = m_caster->GetUInt32Value(UNIT_FIELD_MAXPOWER1);
 		if(cm <= mm)
@@ -6659,7 +6659,7 @@ void Aura::EventPeriodicBurn(uint32 amount, uint32 misc)
 
 		uint32 field = UNIT_FIELD_POWER1 + misc;
 	
-		uint32 Amount = min( amount, m_target->GetUInt32Value( field ) );
+		uint32 Amount = (uint32)min( amount, m_target->GetUInt32Value( field ) );
 		uint32 newHealth = m_target->GetUInt32Value(field) - Amount ;
 				
 		SendPeriodicAuraLog(m_target, m_target, m_spellProto->Id, m_spellProto->School, newHealth, 0, 0, FLAG_PERIODIC_DAMAGE);
